+++ front/src/app/features/features.component.ts
@@ -30,6 +30,7 @@ interface FeatureCategory {
         </a>
         <div class="features-nav__links">
           <a routerLink="/features" class="features-nav__link features-nav__link--active">{{ 'LANDING.NAV_FEATURES' | translate }}</a>
+          <a routerLink="/pricing" class="features-nav__link">{{ 'LANDING.NAV_PRICING' | translate }}</a>
           <a routerLink="/" fragment="guests" class="features-nav__link">{{ 'LANDING.NAV_GUESTS' | translate }}</a>
           <a routerLink="/" fragment="demo" class="features-nav__link">{{ 'LANDING.NAV_DEMO' | translate }}</a>
         </div>